reorganizing translation code to ease header inclusion and compilation
