+++ src/matUtils/extract.hpp
@@ -4,6 +4,7 @@
 #include "describe.hpp"
 #include "uncertainty.hpp"
 #include "select.hpp"
+#include "translate.hpp"
 
 po::variables_map parse_extract_command(po::parsed_options parsed);
 void extract_main (po::parsed_options parsed);
